add xml config parser
I'm getting ready for supporting app delivery from multiple platforms,
and the easiest way to do that will be to start with some scripts.
In preparation for that, I'm adding support for jsonp callbacks that
can take an XML config and turn it into json that can then be used
to manually configure an app on-page. I'm envisioning modifying
the DOM of the Moodle settings page to show the app picker, then
populating the form correctly based on what the user selects.

test plan:
- make sure the tool can process xml
- make sure the tool does some lightweight validation
- make sure jsonp responses are working

Change-Id: I728270b0291a06affae9fc90f91433c1912ef5c0
Reviewed-on: https://gerrit.instructure.com/19345
Reviewed-by: Brian Whitmer <brian@instructure.com>
Product-Review: Brian Whitmer <brian@instructure.com>
QA-Review: Brian Whitmer <brian@instructure.com>
Tested-by: Brian Whitmer <brian@instructure.com>